+++ src/bootstrap/galaxy.ts
@@ -31,16 +31,15 @@ requestAnimationFrame(() => {
 
     require('../components/nvim/statusline')
     require('../components/nvim/command-line')
-    require('../components/experimental/vim-search')
+    require('../components/nvim/vim-search')
   })
 
   setTimeout(() => {
     // TODO(smolck): `requireDir` doesn't recursively search through directories
     // for files, right?
-    requireDir(`${__dirname}/../components/nvim`)
-    requireDir(`${__dirname}/../components/lsp`)
     requireDir(`${__dirname}/../components`)
-    requireDir(`${__dirname}/../components/experimental`)
+    requireDir(`${__dirname}/../components/nvim`)
+    requireDir(`${__dirname}/../components/extensions`)
     requireDir(`${__dirname}/../components/memes`)
   }, 600)
