+++ src/bootstrap/index.html
@@ -62,7 +62,7 @@
       }
 
       /* Without this, there's extra margin on the top and bottom of the
-         hover (src/components/hover.ts) */
+         hover (src/components/extensions/lsp-hover.ts) */
       p,
       pre {
         margin: 0;
