+++ src/components/extensions/future/vim-rename.ts
