+++ src/components/extensions/generic-menu.ts
@@ -1,8 +1,8 @@
-import { Plugin } from '../components/core/plugin-container'
-import { RowNormal } from '../components/core/row-container'
-import { h, app, vimBlur, vimFocus } from '../ui/uikit'
-import { CreateTask } from '../support/utils'
-import Input from '../components/core/text-input'
+import { Plugin } from '../plugin-container'
+import { RowNormal } from '../row-container'
+import { h, app, vimBlur, vimFocus } from '../../ui/uikit'
+import { CreateTask } from '../../support/utils'
+import Input from '../text-input'
 import { filter } from 'fuzzaldrin-plus'
 import * as Icon from 'hyperapp-feather'
 import { Component } from 'hyperapp'
