+++ src/components/extensions/legacy/buffer-search.ts
