+++ src/components/extensions/legacy/debug.ts
