+++ src/components/extensions/legacy/divination.ts
