+++ src/components/extensions/legacy/inventory-search.ts
