+++ src/components/extensions/legacy/inventory.ts
