+++ src/components/extensions/legacy/problem-info.ts
