+++ src/components/extensions/legacy/spell-check.ts
