+++ src/components/extensions/lsp-code-action.ts
@@ -1,8 +1,8 @@
-import { RowNormal } from '../core/row-container'
+import { RowNormal } from '../row-container'
 import { h, app, vimBlur, vimFocus } from '../../ui/uikit'
 import * as windows from '../../windows/window-manager'
-import Input from '../core/text-input'
-import Overlay from '../core/overlay'
+import Input from '../text-input'
+import Overlay from '../overlay'
 import { filter } from 'fuzzaldrin-plus'
 import * as Icon from 'hyperapp-feather'
 import api from '../../core/instance-api'
