+++ src/components/extensions/lsp-hover.ts
@@ -2,7 +2,7 @@ import * as windows from '../../windows/window-manager'
 import { ColorData } from '../../services/colorizer'
 import { sub } from '../../messaging/dispatch'
 import { debounce } from '../../support/utils'
-import Overlay from '../core/overlay'
+import Overlay from '../overlay'
 import { docStyle } from '../../ui/styles'
 import { cursor } from '../../core/cursor'
 import { h, app } from '../../ui/uikit'
