+++ src/components/nvim/autocomplete.tsx
@@ -1,15 +1,15 @@
-import { RowNormal, RowComplete } from '../core/row-container'
+import { RowNormal, RowComplete } from '../row-container'
 import { resetMarkdownHTMLStyle } from '../../ui/styles'
 import * as windows from '../../windows/window-manager'
 import * as dispatch from '../../messaging/dispatch'
 import * as workspace from '../../core/workspace'
 import { PopupMenu } from '../../render/events'
 import { paddingVH, cvar } from '../../ui/css'
-import Overlay from '../core/overlay'
+import Overlay from '../overlay'
 import { cursor } from '../../core/cursor'
 import { parse as stringToMarkdown } from 'marked'
 import { render } from 'inferno'
-import Icon from '../core/icon'
+import Icon from '../icon'
 const feather = require('feather-icons')
 
 export interface CompletionShow {
