+++ src/components/nvim/message-history.ts
@@ -1,7 +1,7 @@
 import { MessageKind, Message } from '../../protocols/veonim'
-import { RowNormal } from '../core/row-container'
+import { RowNormal } from '../row-container'
 import { h, app, vimBlur, vimFocus } from '../../ui/uikit'
-import Input from '../core/text-input'
+import Input from '../text-input'
 import { filter } from 'fuzzaldrin-plus'
 import * as Icon from 'hyperapp-feather'
 import { colors } from '../../ui/styles'
