+++ src/components/plugin-container.tsx
