+++ src/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "exclude": ["node_modules", "test", "xdg_config", "components/legacy", "dev"],
+  "exclude": ["node_modules", "test", "xdg_config", "components/extensions/legacy", "dev"],
   "compilerOptions": {
     "skipLibCheck": true,
     "target": "esnext",
